move the reading of text into a separate file
Will cleanup even more later.
